// data/github/current/ui/packages/item-picker2/projects/stories/__generated__/ItemPickerProjectsBoxStoryQuery.graphql.ts
export type ItemPickerProjectsBoxStoryQuery$variables = {
  number: number;
  owner: string;
  repo: string;
};
export type ItemPickerProjectsBoxStoryQuery$data = {
  readonly repository: {
    readonly issue: {
      readonly projectCards: {
        readonly " $fragmentSpreads": FragmentRefs<"ItemPickerProjectsBox_SelectedClassicProjectCardsFragment">;
      };
      readonly projectsV2: {
        readonly " $fragmentSpreads": FragmentRefs<"ItemPickerProjectsBox_SelectedProjectsV2Fragment">;
      };
    } | null | undefined;
  } | null | undefined;
};
export type ItemPickerProjectsBoxStoryQuery = {
  response: ItemPickerProjectsBoxStoryQuery$data;
  variables: ItemPickerProjectsBoxStoryQuery$variables;
};